Declare ACL mode as a ULONG
Since it is a bitmap of uint32 in the parser (see parsenodes.h), it
ought to be declared as a ULONG in ORCA (corresponds to uint32_t).

This also rids us of the compiler WARNING:

CTranslatorDXLToPlStmt.cpp:5215:42: warning: comparison of integer expressions of different signedness: ‘gpos::INT’ {aka ‘int’} and ‘unsigned int’ [-Wsign-compare]
 5215 |                                 acl_mode <= std::numeric_limits<AclMode>::max());
      |                                 ~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/backend/gpopt/translate/CTranslatorDXLToPlStmt.cpp
@@ -5210,7 +5210,7 @@ CTranslatorDXLToPlStmt::ProcessDXLTblDescr(
 		(void) base_table_context->InsertMapping(dxl_col_descr->Id(), attno);
 	}
 
-	INT acl_mode = table_descr->GetAclMode();
+	ULONG acl_mode = table_descr->GetAclMode();
 	GPOS_ASSERT(acl_mode >= 0 &&
 				acl_mode <= std::numeric_limits<AclMode>::max());
 	AclMode required_perms = static_cast<AclMode>(acl_mode);
